refactor: response, error 관련된 코드 global 폴더로 이동
